Modify homework after review

+++ app/controllers/api/bookings_controller.rb
@@ -7,9 +7,13 @@ module Api
 
     # GET /api/bookings/:id
     def show
-      booking = Booking.find(params[:id])
+      booking = Booking.find_by(id: params[:id])
 
-      render json: BookingSerializer.render(booking, root: :booking)
+      if booking
+        render json: BookingSerializer.render(booking, root: :booking)
+      else
+        render json: { errors: "Couldn't find the Booking" }, status: :bad_request
+      end
     end
 
     # POST /api/bookings
@@ -25,26 +29,35 @@ module Api
 
     # DELETE /api/bookings/:id
     def destroy
-      booking = Booking.find(params[:id])
-
-      booking.destroy
+      booking = Booking.find_by(id: params[:id])
+      if booking
+        booking.destroy
+      else
+        render json: { errors: "Couldn't find Booking" }, status: :bad_request
+      end
     end
 
     # PATCH /api/bookings/:id
     def update
-      booking = Booking.find(params[:id])
+      booking = Booking.find_by(id: params[:id])
 
-      if booking.update(booking_params)
-        render json: BookingSerializer.render(booking, root: :booking)
-      else
-        render json: { errors: booking.errors }, status: :bad_request
-      end
+      return booking_update(booking) if booking
+
+      render json: { errors: "Couldn't find the Booking" }, status: :bad_request
     end
 
     private
 
     def booking_params
       params.require(:booking).permit(:no_of_seats, :seat_price, :user_id, :flight_id)
+    end
+
+    def booking_update(booking)
+      if booking.update(booking_params)
+        render json: BookingSerializer.render(booking, root: :booking)
+      else
+        render json: { errors: booking.errors }, status: :bad_request
+      end
     end
   end
 end
